requirements.txt
CacheControl==0.12.11
cachelib==0.9.0
cachetools==5.2.0
certifi==2022.9.24
cffi==1.15.1
charset-normalizer==2.1.1
click==8.1.3
cryptography==38.0.1
Flask==2.2.2
Flask-Cors==3.0.10
Flask-Session==0.4.0
grpcio==1.49.1
grpcio-status==1.49.1
httplib2==0.20.4
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
msgpack==1.0.4
oauthlib==3.2.1
proto-plus==1.22.1
protobuf==4.21.7
psycopg2-binary==2.9.5
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
pyparsing==3.0.9
python-dotenv==0.21.0
requests==2.28.1
requests-oauthlib==1.3.1
rsa==4.9
six==1.16.0
tweepy==4.10.1
uritemplate==4.1.1
urllib3==1.26.12
waitress==2.1.2
Werkzeug==2.2.2
